pyproject.toml
[tool.poetry]
name = "plaudpy"
version = "0.1.0"
description = "Python library for interfacing with the Plaud.ai API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "plaudpy", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27"
pydantic = "^2.0"
pydantic-settings = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "functional: marks tests as functional (require PLAUD_USERNAME and PLAUD_PASSWORD)",
]
